Lab8 started

+++ l6/lab6_3.m
@@ -12,7 +12,7 @@ M = [ 0  1  0;
       0  1  0];
 M = M/9;
 
-convoluted = uint8(conv2(moon, laplacianFilter, 'same'));
+convoluted = conv2(moon, laplacianFilter, 'same');
 
 figure(1);
 subplot(2, 2, 1);
@@ -20,7 +20,7 @@ imshow(moon, []);
 title('original');
 
 subplot(2, 2, 2);
-imshow(convoluted + 128, []);
+imshow(uint8(convoluted + 128), []);
 title('laplacian + 128');
 
 subplot(2, 2, 3);
@@ -28,7 +28,7 @@ imshow(abs(convoluted), []);
 title('laplacian i modul wartosci');
 
 subplot(2, 2, 4);
-imshow(moon + convoluted + 128, []);
+imshow(moon + uint8(convoluted + 128), []);
 title('moon + 128');
